ultima entrega recomendada

+++ cliente/misfunciones.c
@@ -246,9 +246,9 @@ int initsocket(struct addrinfo *servinfo, char f_verbose)
 /**************************************************************************/
 /* Funcion envio de mensaje */
 /**************************************************************************/
-ssize_t enviarDatos(struct rcftp_msg *mensaje_enviar, int socket, struct sockaddr *remote, socklen_t remotelen)
+int enviarDatos(struct rcftp_msg *mensaje_enviar, int socket, struct sockaddr *remote, socklen_t remotelen)
 {
-	ssize_t sendsize = sendto(socket, mensaje_enviar, sizeof(*mensaje_enviar), 0, remote, remotelen);
+	int sendsize = sendto(socket, mensaje_enviar, sizeof(*mensaje_enviar), 0, remote, remotelen);
 	if (sendsize != sizeof(*mensaje_enviar) || sendsize < 0)
 	{
 		printf("Error en sendto");
@@ -259,9 +259,9 @@ ssize_t enviarDatos(struct rcftp_msg *mensaje_enviar, int socket, struct sockadd
 /**************************************************************************/
 /*  Función para recibir datos  */
 /**************************************************************************/
-ssize_t recibirDatos(int socket, struct rcftp_msg *buffer, int length, struct addrinfo *servinfo)
+int recibirDatos(int socket, struct rcftp_msg *buffer, int length, struct addrinfo *servinfo)
 {
-	ssize_t recvsize = recvfrom(socket, (char *)buffer, length, 0, (struct sockaddr *)servinfo, &servinfo->ai_addrlen);
+	int recvsize = recvfrom(socket, (char *)buffer, length, 0, (struct sockaddr *)servinfo, &servinfo->ai_addrlen);
 	if (errno != EAGAIN && recvsize < 0)
 	{
 		printf("Error en rcvfrom");
